services/worker/requirements.txt
oracledb==2.0.1
psycopg[binary]==3.1.16
pymongo==4.6.0
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.0
python-json-logger==2.0.7

